# perf: Improve performance of literal de-duplication