@bclement-ocp 2024-07-25
> On our internal dataset, the full patch is a speedup of about 3-4%.

And also a tiny performance boost with a +13-5 delta (absolute +8 problems solved) due to less timeouts.
(The -5 can be explained by the fact that we `Hashtbl` iteration and `Set` iterations are generally not going to follow the same order, and so we process literals in a slightly different order)